UI : UploadScreen TopBar, BottomBar Apply in column - #19
